@@ -30,23 +30,24 @@ use Symfony\Component\Routing\Annotation\Route;
 /**
  * Controller used to manage blog contents in the public part of the site.
  *
- * @Route("/blog")
- *
  * @author Ryan Weaver <weaverryan@gmail.com>
  * @author Javier Eguiluz <javier.eguiluz@gmail.com>
  */
+#[Route('/blog')]
 class BlogController extends AbstractController
 {
     /**
-     * @Route("/", defaults={"page": "1", "_format"="html"}, methods="GET", name="blog_index")
-     * @Route("/rss.xml", defaults={"page": "1", "_format"="xml"}, methods="GET", name="blog_rss")
-     * @Route("/page/{page<[1-9]\d*>}", defaults={"_format"="html"}, methods="GET", name="blog_index_paginated")
-     * @Cache(smaxage="10")
-     *
      * NOTE: For standard formats, Symfony will also automatically choose the best
      * Content-Type header for the response.
+     *
      * See https://symfony.com/doc/current/routing.html#special-parameters
      */
+    #[
+        Route('/', defaults: ['page' => '1', '_format' => 'html'], methods: ['GET'], name: 'blog_index'),
+        Route('/rss.xml', defaults: ['page' => '1', '_format' => 'xml'], methods: ['GET'], name: 'blog_rss'),
+        Route('/page/{page<[1-9]\d*>}', defaults: ['_format' => 'html'], methods: ['GET'], name: 'blog_index_paginated'),
+    ]
+    #[Cache(smaxage: 10)]
     public function index(Request $request, int $page, string $_format, PostRepository $posts, TagRepository $tags): Response
     {
         $tag = null;
@@ -65,13 +66,13 @@ class BlogController extends AbstractController
     }
 
     /**
-     * @Route("/posts/{slug}", methods="GET", name="blog_post")
-     *
      * NOTE: The $post controller argument is automatically injected by Symfony
      * after performing a database query looking for a Post with the 'slug'
      * value given in the route.
+     *
      * See https://symfony.com/doc/current/bundles/SensioFrameworkExtraBundle/annotations/converters.html
      */
+    #[Route('/posts/{slug}', methods: ['GET'], name: 'blog_post')]
     public function postShow(Post $post): Response
     {
         // Symfony's 'dump()' function is an improved version of PHP's 'var_dump()' but
@@ -85,15 +86,15 @@ class BlogController extends AbstractController
     }
 
     /**
-     * @Route("/comment/{postSlug}/new", methods="POST", name="comment_new")
-     * @IsGranted("IS_AUTHENTICATED_FULLY")
-     * @ParamConverter("post", options={"mapping": {"postSlug": "slug"}})
-     *
      * NOTE: The ParamConverter mapping is required because the route parameter
      * (postSlug) doesn't match any of the Doctrine entity properties (slug).
+     *
      * See https://symfony.com/doc/current/bundles/SensioFrameworkExtraBundle/annotations/converters.html#doctrine-converter
      */
-    public function commentNew(Request $request, Post $post, EventDispatcherInterface $eventDispatcher, EntityManagerInterface $entityManager): Response
+    #[Route('/comment/{postSlug}/new', methods: ['POST'], name: 'comment_new')]
+    #[IsGranted('IS_AUTHENTICATED_FULLY')]
+    #[ParamConverter('post', options: ['mapping' => ['postSlug' => 'slug']])]
+    public function commentNew(Request $request, Post $post, EventDispatcherInterface $eventDispatcher,  EntityManagerInterface $entityManager): Response
     {
         $comment = new Comment();
         $comment->setAuthor($this->getUser());
@@ -140,9 +141,7 @@ class BlogController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/search", methods="GET", name="blog_search")
-     */
+    #[Route('/search', methods: ['GET'], name: 'blog_search')]
     public function search(Request $request, PostRepository $posts): Response
     {
         $query = $request->query->get('q', '');
