@@ -16,11 +16,9 @@ use Symfony\Contracts\EventDispatcher\Event;
 
 class CommentCreatedEvent extends Event
 {
-    protected $comment;
-
-    public function __construct(Comment $comment)
-    {
-        $this->comment = $comment;
+    public function __construct(
+        protected Comment $comment
+    ) {
     }
 
     public function getComment(): Comment
