@@ -25,17 +25,12 @@ use Symfony\Contracts\Translation\TranslatorInterface;
  */
 class CommentNotificationSubscriber implements EventSubscriberInterface
 {
-    private $mailer;
-    private $translator;
-    private $urlGenerator;
-    private $sender;
-
-    public function __construct(MailerInterface $mailer, UrlGeneratorInterface $urlGenerator, TranslatorInterface $translator, string $sender)
-    {
-        $this->mailer = $mailer;
-        $this->urlGenerator = $urlGenerator;
-        $this->translator = $translator;
-        $this->sender = $sender;
+    public function __construct(
+        private MailerInterface $mailer,
+        private UrlGeneratorInterface $urlGenerator,
+        private TranslatorInterface $translator,
+        private string $sender
+    ) {
     }
 
     public static function getSubscribedEvents(): array
