@@ -28,14 +28,14 @@ use function Symfony\Component\String\u;
  */
 class RedirectToPreferredLocaleSubscriber implements EventSubscriberInterface
 {
-    private $urlGenerator;
-    private $locales;
-    private $defaultLocale;
+    private array $locales;
+    private string $defaultLocale;
 
-    public function __construct(UrlGeneratorInterface $urlGenerator, string $locales, string $defaultLocale = null)
-    {
-        $this->urlGenerator = $urlGenerator;
-
+    public function __construct(
+        private UrlGeneratorInterface $urlGenerator,
+        string $locales,
+        ?string $defaultLocale = null
+    ) {
         $this->locales = explode('|', trim($locales));
         if (empty($this->locales)) {
             throw new \UnexpectedValueException('The list of supported locales must not be empty.');
