@@ -27,11 +27,9 @@ use function Symfony\Component\String\u;
  */
 class TagArrayToStringTransformer implements DataTransformerInterface
 {
-    private $tags;
-
-    public function __construct(TagRepository $tags)
-    {
-        $this->tags = $tags;
+    public function __construct(
+        private TagRepository $tags
+    ) {
     }
 
     /**
