@@ -31,12 +31,10 @@ use Symfony\Component\String\Slugger\SluggerInterface;
  */
 class PostType extends AbstractType
 {
-    private $slugger;
-
     // Form types are services, so you can inject other services in them if needed
-    public function __construct(SluggerInterface $slugger)
-    {
-        $this->slugger = $slugger;
+    public function __construct(
+        private SluggerInterface $slugger
+    ) {
     }
 
     /**
