@@ -22,10 +22,8 @@ class MomentFormatConverter
      * This defines the mapping between PHP ICU date format (key) and moment.js date format (value)
      * For ICU formats see http://userguide.icu-project.org/formatparse/datetime#TOC-Date-Time-Format-Syntax
      * For Moment formats see https://momentjs.com/docs/#/displaying/format/.
-     *
-     * @var array
      */
-    private static $formatConvertRules = [
+    private const FORMAT_CONVERT_RULES = [
         // year
         'yyyy' => 'YYYY', 'yy' => 'YY', 'y' => 'YYYY',
         // day
@@ -43,6 +41,6 @@ class MomentFormatConverter
      */
     public function convert(string $format): string
     {
-        return strtr($format, self::$formatConvertRules);
+        return strtr($format, self::FORMAT_CONVERT_RULES);
     }
 }
